# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Todo List Service API
  description: RESTful API for managing a simple to-do list
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: todos
    description: Todo item management

paths:
  /api/v1/todos:
    post:
      tags:
        - todos
      summary: Create a new todo item
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - todos
      summary: Get todo items (defaults to not done only)
      operationId: getTodos
      parameters:
        - name: status
          in: query
          description: Use "all" to get all items regardless of status. Defaults to not done items only.
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Zero-based page index.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of items per page.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        '200':
          description: Paginated list of todo items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoPageResponse'

  /api/v1/todos/{id}:
    get:
      tags:
        - todos
      summary: Get a single todo item by ID
      operationId: getTodoById
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          description: Todo item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - todos
      summary: Update a todo item description
      operationId: updateTodoDescription
      parameters:
        - $ref: '#/components/parameters/TodoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Item is past due and cannot be modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/todos/{id}/done:
    patch:
      tags:
        - todos
      summary: Mark a todo item as done
      operationId: markTodoDone
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          description: Todo item marked as done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Item is past due and cannot be modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/todos/{id}/not-done:
    patch:
      tags:
        - todos
      summary: Mark a todo item as not done
      operationId: markTodoNotDone
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          description: Todo item marked as not done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Item is past due and cannot be modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    TodoId:
      name: id
      in: path
      required: true
      description: Unique identifier of the todo item
      schema:
        type: string
        format: uuid

  schemas:
    CreateTodoRequest:
      type: object
      required:
        - description
        - dueDatetime
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Text detailing the task
          example: Buy groceries
        dueDatetime:
          type: string
          format: date-time
          description: Deadline for the task (must be in the future)
          example: '2026-03-15T10:00:00Z'

    UpdateTodoRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Updated text for the task
          example: Buy organic groceries

    TodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        description:
          type: string
          description: Text detailing the task
          example: Buy groceries
        status:
          type: string
          description: Current status of the item
          enum:
            - NOT_DONE
            - DONE
            - PAST_DUE
          example: NOT_DONE
        creationDatetime:
          type: string
          format: date-time
          description: When the item was created
          example: '2026-02-12T08:00:00Z'
        dueDatetime:
          type: string
          format: date-time
          description: Deadline for the task
          example: '2026-03-15T10:00:00Z'
        doneDatetime:
          type: string
          format: date-time
          nullable: true
          description: When the item was marked as done (null if not done)

    TodoPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoResponse'
        totalCount:
          type: integer
          format: int64
          description: Total number of items matching the query

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: HTTP status reason phrase
        message:
          type: string
          description: Human-readable error message